refactor(runner): merge api client into provider module
Inline ApiClient (135 lines) into provider/api.rs as a private struct.
Remove src/api.rs — the HTTP transport is now fully encapsulated inside
the provider module with no external consumers.
