+++ crates/runner/src/main.rs
@@ -1,4 +1,3 @@
-mod api;
 mod cmd;
 mod config;
 mod deps;
